requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.2
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
packaging==25.0
pillow==11.0.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
pytesseract==0.3.13
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.3
tokenizers==0.15.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.35.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
